everething is refactored. joke. SVNDiffWindow now keeps plain instructions and new data bytes inside, for API users that means that ISVNEditor.textDeltaChunk(...) may return "null" now, as returned stream is no longer used.

+++ javasvn/src/org/tmatesoft/svn/core/io/diff/SVNDiffWindowApplyBaton.java
@@ -35,9 +35,9 @@ public class SVNDiffWindowApplyBaton {
     InputStream mySourceStream;
     OutputStream myTargetStream;
 
-    long mySourceViewOffset;
-    long mySourceViewLength;
-    long myTargetViewSize;
+    int mySourceViewOffset;
+    int mySourceViewLength;
+    int myTargetViewSize;
 
     byte[] mySourceBuffer;
     byte[] myTargetBuffer;
